Aero changes
Now the program computes the AoA in each component and calculates its CN (AoA changes along the length due to the pitching velocity).
There is no longer juggling to get the CN of the control fins, now everything is done properly and the 3D arrows are computed separately.
Green arrow now includes ONLY the body and stabilization fins, added to the red one of the control fins gives the total force acting on the rocket (i.e. the green arrow is not located in the real CP, but on the "passive" CP)
Added CN, CA and CM to the plots.