Aero changes
Made dynamic pressure scaling optional (in the source code, it is disabled by default). Now rockets should tumble more, not sure if it's more realistic or not, but it should be more conservative.

Modified the Barrowman CN for better results at AoA > 45º

Added some optional experimental features in the source code, like drag based on the component Reynolds or a better discretization of the rocket (more resource intensive).